refactored util package to text package and wasm main func
