+++ pkg/text/text.go
@@ -1,4 +1,4 @@
-package util
+package text
 
 import "unicode"
