+++ pkg/text/text_test.go
@@ -1,4 +1,4 @@
-package util
+package text
 
 import "testing"
